Open videos as unsaved documents
This means we no longer need to override fileURL (because it is actually
nil).

Found a nice way to make the document re-opening behavior automatic
again.

I was looking to make state restoration / autosaving work for me here,
but I cannot get NSDocument from throwing up the save dialog (unless
I reset the change-count to 0), and even then I don't know how
restoration on start-up is supposed to be triggered.